Removed redundant variables;

+++ src/terra.sh
@@ -68,17 +68,20 @@ function sc_terra_versions() {
 # Up
 ########################################################################################################################
 function sc_terra_up_init() {
-    sc_heading 1 "Initializing terraform"
-    if [[ -n "$_ARG_INIT" ]]; then
-        rm -rf "${_ST_CONTEXT_HOME}/"{terraform*,.terraform,modules/bootstrap/assets}
+    if exo compute sks versions --output-format text | grep -Eq "^${_ST_VERSION_KUBERNETES}$"; then
+        sc_heading 1 "Initializing terraform"
+        if [[ -n "$_ARG_INIT" ]]; then
+            rm -rf "${_ST_CONTEXT_HOME}/"{terraform*,.terraform,modules/bootstrap/assets}
+        fi
+        terraform -chdir="$_ST_CONTEXT_HOME" init -upgrade=true
+        sc_terra_versions
+    else
+        echo "Kubernetes version $_ST_VERSION_KUBERNETES is not available. Aborting..."
+        exit 1
     fi
-    terraform -chdir="$_ST_CONTEXT_HOME" init -upgrade=true
-    sc_terra_versions
 }
 
 function sc_terra_up_assets() {
-    sc_terra_down_assets
-    sc_terra_down_bucket
     sc_heading 2 "Creating assets..."
     terraform -chdir="$_ST_CONTEXT_HOME" apply \
         -auto-approve \
@@ -89,7 +92,7 @@ function sc_terra_up_assets() {
         -var="cluster_name=${_ST_STAGE}"
 }
 
-function sc_terra_up_iam() {
+function sc_terra_up_iam_backup() {
     for _role in backup data scaler traces; do
         echo -n "Saving credentials for ${_role}..."
         local _file="${_ST_CONTEXT_HOME}/${_role}.iam"
@@ -108,51 +111,43 @@ function sc_terra_up_iam() {
 }
 
 function sc_terra_up() {
-    if exo compute sks versions --output-format text | grep -Eq "^${_ST_VERSION_KUBERNETES}$"; then
-        sc_terra_up_init
+    sc_terra_up_init
 
-        if [[ -n "$_ARG_DRYRUN" ]]; then
-            local -r _serenditree_tf="${_ST_CONTEXT_HOME}/serenditree.tf"
-            # shellcheck disable=SC2064
-            trap "mv ${_serenditree_tf}.bak ${_serenditree_tf}" EXIT
-            sed -Ei.bak 's/set_sensitive/set/' "${_serenditree_tf}"
-        fi
+    if [[ -n "$_ARG_DRYRUN" ]]; then
+        local -r _serenditree_tf="${_ST_CONTEXT_HOME}/serenditree.tf"
+        # shellcheck disable=SC2064
+        trap "mv ${_serenditree_tf}.bak ${_serenditree_tf}" EXIT
+        sed -Ei.bak 's/set_sensitive/set/' "${_serenditree_tf}"
+    fi
 
-        sc_heading 1 "Planing terraform"
-        local -r _plan='serenditree.tfplan'
-        terraform -chdir="$_ST_CONTEXT_HOME" plan \
-            -var=api_key="$(pass serenditree/iam/serenditree@exoscale.com.access)" \
-            -var=api_secret="$(pass serenditree/iam/serenditree@exoscale.com.secret)" \
-            -var=kubernetes_version="${_ST_VERSION_KUBERNETES}" \
-            -var=account="${_ST_ACCOUNT}" \
-            -var=context="${_ST_CONTEXT}" \
-            -var=cluster_domain="cluster.local" \
-            -var=stage="${_ST_STAGE}" \
-            -var=host="${_ST_DOMAIN}" \
-            -var=issuer="${_ARG_ISSUER}" \
-            -var=zone="${_ST_ZONE}" \
-            -var=git_repo="${_ST_GIT}" \
-            -var=git_ssh="${_ST_GIT_SSH}" \
-            -var=app_parameters="$(sc_terra_secrets app)" \
-            -var=cicd_parameters="$(sc_terra_secrets cicd)" \
-            -var=o11y_parameters="$(sc_terra_secrets o11y)" \
-            -var=oidc_parameters="$(sc_terra_secrets oidc)" \
-            -out "$_plan"
+    sc_heading 1 "Planing terraform"
+    local -r _plan='serenditree.tfplan'
+    terraform -chdir="$_ST_CONTEXT_HOME" plan \
+        -var=api_key="$(pass serenditree/iam/serenditree@exoscale.com.access)" \
+        -var=api_secret="$(pass serenditree/iam/serenditree@exoscale.com.secret)" \
+        -var=kubernetes_version="${_ST_VERSION_KUBERNETES}" \
+        -var=account="${_ST_ACCOUNT}" \
+        -var=context="${_ST_CONTEXT}" \
+        -var=stage="${_ST_STAGE}" \
+        -var=host="${_ST_DOMAIN}" \
+        -var=issuer="${_ARG_ISSUER}" \
+        -var=zone="${_ST_ZONE}" \
+        -var=app_parameters="$(sc_terra_secrets app)" \
+        -var=cicd_parameters="$(sc_terra_secrets cicd)" \
+        -var=o11y_parameters="$(sc_terra_secrets o11y)" \
+        -var=oidc_parameters="$(sc_terra_secrets oidc)" \
+        -out "$_plan"
 
-        if [[ -n "$_ARG_DRYRUN" ]]; then
-            sc_terra_render "$_plan"
-        else
-            sc_heading 1 "Applying terraform"
-            terraform -chdir="$_ST_CONTEXT_HOME" apply "$_plan"
-            sc_heading 1 "Backing up IAM credentials"
-            sc_terra_up_iam
-            sc_heading 1 "Setting up context"
-            sc_context_clean
-            sc_context_init_kube "${_ST_CONTEXT_HOME}/kubeconfig"
-        fi
+    if [[ -n "$_ARG_DRYRUN" ]]; then
+        sc_terra_render "$_plan"
     else
-        echo "Kubernetes version $_ST_VERSION_KUBERNETES is not available. Aborting..."
-        exit 1
+        sc_heading 1 "Applying terraform"
+        terraform -chdir="$_ST_CONTEXT_HOME" apply "$_plan"
+        sc_heading 1 "Backing up IAM credentials"
+        sc_terra_up_iam_backup
+        sc_heading 1 "Setting up context"
+        sc_context_clean
+        sc_context_init_kube "${_ST_CONTEXT_HOME}/kubeconfig"
     fi
 }
 ########################################################################################################################
@@ -176,57 +171,32 @@ function sc_terra_down_volumes() {
     xargs -n1 exo compute block-storage delete --force </tmp/serenditree-pv
 }
 
-function sc_terra_down_bucket() {
-    echo "Removing bucket..."
-    if exo storage show sos://okd 2>/dev/null; then
-        exo storage rb sos://okd --recursive --force
-    else
-        echo "error: bucket does not exits."
-    fi
-}
-
-function sc_terra_down_assets() {
-    echo "Removing assets..."
-    rm -rfv "${_ST_CONTEXT_HOME}/modules/bootstrap/assets/${_ST_STAGE}"
-}
-
 function sc_terra_down() {
-    if [[ -z "$_ARG_DRYRUN" ]]; then
-        sc_heading 2 "Cluster volumes:"
-        kubectl get pv --output=custom-columns='name:.metadata.name' --no-headers | tee /tmp/serenditree-pv
+    sc_heading 2 "Cluster volumes:"
+    kubectl get pv --output=custom-columns='name:.metadata.name' --no-headers | tee /tmp/serenditree-pv
 
-        sc_heading 2 "Removing helm releases from state..."
-        terraform -chdir="$_ST_CONTEXT_HOME" show -json serenditree.tfplan |
-            jq -r ".configuration.root_module.resources[] | select(.type == \"helm_release\") | .address" |
-            xargs terraform -chdir="$_ST_CONTEXT_HOME" state rm
+    sc_heading 2 "Removing helm releases from state..."
+    terraform -chdir="$_ST_CONTEXT_HOME" show -json serenditree.tfplan |
+        jq -r ".configuration.root_module.resources[] | select(.type == \"helm_release\") | .address" |
+        xargs terraform -chdir="$_ST_CONTEXT_HOME" state rm
 
-        sc_heading 2 "Starting deletion..."
-        terraform -chdir="$_ST_CONTEXT_HOME" destroy -auto-approve \
-            -var=api_key="$(pass serenditree/iam/serenditree@exoscale.com.access)" \
-            -var=api_secret="$(pass serenditree/iam/serenditree@exoscale.com.secret)" \
-            -var=kubernetes_version="${_ST_VERSION_KUBERNETES}" \
-            -var=account="${_ST_ACCOUNT}" \
-            -var=context="${_ST_CONTEXT}" \
-            -var=cluster_domain="cluster.local" \
-            -var=stage="${_ST_STAGE}" \
-            -var=host="${_ST_DOMAIN}" \
-            -var=issuer="${_ARG_ISSUER}" \
-            -var=zone="${_ST_ZONE}" \
-            -var=git_repo="${_ST_GIT}" \
-            -var=git_ssh="${_ST_GIT_SSH}" \
-            -var=app_parameters="$(sc_terra_secrets app)" \
-            -var=cicd_parameters="$(sc_terra_secrets cicd)" \
-            -var=o11y_parameters="$(sc_terra_secrets o11y)" \
-            -var=oidc_parameters="$(sc_terra_secrets oidc)"
+    sc_heading 2 "Starting deletion..."
+    terraform -chdir="$_ST_CONTEXT_HOME" destroy -auto-approve \
+        -var=api_key="$(pass serenditree/iam/serenditree@exoscale.com.access)" \
+        -var=api_secret="$(pass serenditree/iam/serenditree@exoscale.com.secret)" \
+        -var=kubernetes_version="${_ST_VERSION_KUBERNETES}" \
+        -var=account="${_ST_ACCOUNT}" \
+        -var=context="${_ST_CONTEXT}" \
+        -var=stage="${_ST_STAGE}" \
+        -var=host="${_ST_DOMAIN}" \
+        -var=issuer="${_ARG_ISSUER}" \
+        -var=zone="${_ST_ZONE}" \
+        -var=app_parameters="$(sc_terra_secrets app)" \
+        -var=cicd_parameters="$(sc_terra_secrets cicd)" \
+        -var=o11y_parameters="$(sc_terra_secrets o11y)" \
+        -var=oidc_parameters="$(sc_terra_secrets oidc)"
 
-        sc_prompt "Remove DNS records?" sc_terra_down_dns
-        if [[ -n "${_ST_CONTEXT_KUBERNETES}" ]]; then
-            sc_prompt "Delete loadbalancer?" sc_terra_down_loadbalancer
-            sc_prompt "Delete volumes?" sc_terra_down_volumes
-        fi
-        if [[ -n "${_ST_CONTEXT_OPENSHIFT}" ]]; then
-            sc_prompt "Remove bucket?" sc_terra_down_bucket
-            sc_prompt "Remove assets?" sc_terra_down_assets
-        fi
-    fi
+    sc_prompt "Remove DNS records?" sc_terra_down_dns
+    sc_prompt "Delete loadbalancer?" sc_terra_down_loadbalancer
+    sc_prompt "Delete volumes?" sc_terra_down_volumes
 }
